+++ Example/Tests/KFXSectionDataTests.m
@@ -186,6 +186,24 @@
     XCTAssertEqualObjects(cellData1, received1);
 }
 
+-(void)testAllContents{
+    
+    // GIVEN
+    KFXCellData *cellData0 = [KFXCellData cellData];
+    cellData0.contents = @"Content 0";
+    KFXCellData *cellData1 = [KFXCellData cellData];
+    cellData1.contents = @{@"aKey":@"aValue"};
+    [self.sut addCellDataFromArray:@[cellData0,cellData1]];
+    
+    // WHEN
+    NSArray *contents = [self.sut allContents];
+    
+    // THEN
+    XCTAssertNotNil(contents);
+    XCTAssertEqual(2, contents.count);
+    XCTAssertEqualObjects(contents.firstObject, cellData0.contents);
+    XCTAssertEqualObjects(contents.lastObject, cellData1.contents);
+}
 
 //--------------------------------------------------------
 #pragma mark Test Queries